Move the fortran function to its own file and make clean removes version.f90
