+++ Common.mk
@@ -32,7 +32,7 @@ endif
 #      *                                                                *
 #      ******************************************************************
 
-MAKE_CLEAN_ARGUMENTS = *~ *.o *.mod *.il *.stb c_* *.so
+MAKE_CLEAN_ARGUMENTS = *~ *.o *.mod *.il *.stb c_* *.so version.f90
 
 #      ******************************************************************
 #      *                                                                *
