+++ pyspline/__init__.py
@@ -3,3 +3,4 @@ __version__ = "1.5.0"
 from .pyCurve import Curve
 from .pySurface import Surface
 from .pyVolume import Volume
+from .fortranVersion import fortranVersion
